Map struct (#19)
* quoteTestDone

* cryptocurrenciesTest

* operationServiceTestUpdated

* operationTestUpdated

* fixBug

* refactorTests

* fixTest

* operationTestRefactor-IntentionPost

* operationTestRefactor

* limitsAndRangeOf5

* cleanCode

+++ src/main/java/ar/edu/unq/desapp/grupoa022022/backenddesappapi/service/serviceimpl/CryptocurrencyService.java
@@ -1,13 +1,12 @@
 package ar.edu.unq.desapp.grupoa022022.backenddesappapi.service.serviceimpl;
 
+import ar.edu.unq.desapp.grupoa022022.backenddesappapi.dto.CryptocurrencyRegister;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.model.Cryptocurrency;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.model.Intention;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.model.Quote;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.model.exceptions.ResourceNotFound;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.persistence.ICryptocurrencyRepo;
-
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.persistence.IQuoteRepo;
-
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.service.interfaceservice.ICryptocurrencyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,24 +15,29 @@ import java.util.List;
 import java.util.Set;
 
 @Service
-public class CrytocurrencyService implements ICryptocurrencyService {
+public class CryptocurrencyService implements ICryptocurrencyService {
 
     @Autowired
     private ICryptocurrencyRepo cryptocurrencyRepo;
-
 
     @Autowired
     private IQuoteRepo quoteRepo;
 
     @Override
-    public Cryptocurrency create(String name) {
-        Cryptocurrency cryptocurrency = new Cryptocurrency(name);
-        return cryptocurrencyRepo.save(cryptocurrency);
+    public Cryptocurrency create(CryptocurrencyRegister cryptocurrencyRegister) {
+        Cryptocurrency cryptocurrency = new Cryptocurrency(cryptocurrencyRegister.getName());
+        Cryptocurrency savedCryptocurrency = cryptocurrencyRepo.save(cryptocurrency);
+        Quote quote = new Quote(savedCryptocurrency, cryptocurrencyRegister.getPrice());
+        quoteRepo.save(quote);
+        return savedCryptocurrency;
     }
 
     @Override
     public void delete(int id) {
-        cryptocurrencyRepo.deleteById(id);
+        if (cryptocurrencyRepo.findById(id).isPresent()) {
+            cryptocurrencyRepo.findById(id).get().getQuotes().forEach(quote -> quoteRepo.deleteById(quote.getId()));
+            cryptocurrencyRepo.deleteById(id);
+        }
     }
 
     @Override
@@ -43,7 +47,7 @@ public class CrytocurrencyService implements ICryptocurrencyService {
 
     @Override
     public void deleteAll() {
-        cryptocurrencyRepo.deleteAll();
+        cryptocurrencyRepo.findAll().forEach(cryptocurrency -> this.delete(cryptocurrency.getId()));
     }
 
     @Override
